Level1 Half done.
Changed a ton about how the save system works. There are now two forms of data, saveData, and levelData. levelData can be set in editor and is for fixed things like in this example lure info. I've added a class for lure that holds the lures junk chance, cost, and the sprite.